+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.4.3]
+
+### Fixed
+- AirPlay icon does not change into active state
+
 ## [3.4.2]
 
 ### Changed